# fix(resourcemanager): append custom `user-agent` headers