--- a/sdk/resourcemanager/Azure.ResourceManager/Azure.ResourceManager.sln
+++ b/sdk/resourcemanager/Azure.ResourceManager/Azure.ResourceManager.sln
@@ -27,6 +27,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{4E159B41-7
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Core", "..\..\core\Azure.Core\src\Azure.Core.csproj", "{E0F3952A-D1C7-4CE4-B2A8-5C34945EBED8}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.ResourceManager.Perf", "perf\Azure.ResourceManager.Perf.csproj", "{138A2FF1-F1AB-480A-9588-136FB875026E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Test.Perf", "..\..\..\common\Perf\Azure.Test.Perf\Azure.Test.Perf.csproj", "{AD397048-D114-4A32-8F35-64E5C72A4697}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -49,6 +53,14 @@ Global
 		{E0F3952A-D1C7-4CE4-B2A8-5C34945EBED8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E0F3952A-D1C7-4CE4-B2A8-5C34945EBED8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E0F3952A-D1C7-4CE4-B2A8-5C34945EBED8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{138A2FF1-F1AB-480A-9588-136FB875026E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{138A2FF1-F1AB-480A-9588-136FB875026E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{138A2FF1-F1AB-480A-9588-136FB875026E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{138A2FF1-F1AB-480A-9588-136FB875026E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AD397048-D114-4A32-8F35-64E5C72A4697}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD397048-D114-4A32-8F35-64E5C72A4697}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD397048-D114-4A32-8F35-64E5C72A4697}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD397048-D114-4A32-8F35-64E5C72A4697}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE