--- a/sdk/resourcemanager/Azure.ResourceManager/perf/Azure.ResourceManager.Perf.csproj
+++ b/sdk/resourcemanager/Azure.ResourceManager/perf/Azure.ResourceManager.Perf.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <!--Exlucde some shared codes which will cause conflict-->
+    <IsMgmtLibrary>false</IsMgmtLibrary>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\common\Perf\Azure.Test.Perf\Azure.Test.Perf.csproj" />
+    <ProjectReference Include="..\..\..\core\Azure.Core.TestFramework\src\Azure.Core.TestFramework.csproj" />
+  </ItemGroup>
+</Project>