--- a/sdk/resourcemanager/Azure.ResourceManager/src/Internal/MgmtTelemetryPolicy.cs
+++ b/sdk/resourcemanager/Azure.ResourceManager/src/Internal/MgmtTelemetryPolicy.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Runtime.InteropServices;
 using Azure.Core;
 using Azure.Core.Pipeline;
 
@@ -19,6 +18,11 @@ namespace Azure.ResourceManager.Core
         public override void OnSendingRequest(HttpMessage message)
         {
             var header = message.TryGetProperty("UserAgentOverride", out var userAgent) ? userAgent as string : _defaultHeader;
+            if (message.Request.Headers.TryGetValues(HttpHeader.Names.UserAgent, out var customHeaders))
+            {
+                // append custom "user-agent" headers
+                header = $"{header} {string.Join(" ", customHeaders)}";
+            }
             message.Request.Headers.SetValue(HttpHeader.Names.UserAgent, header);
         }
     }