--- a/sdk/resourcemanager/Azure.ResourceManager/src/Internal/MgmtTelemetryPolicy.cs
+++ b/sdk/resourcemanager/Azure.ResourceManager/src/Internal/MgmtTelemetryPolicy.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Text;
 using Azure.Core;
 using Azure.Core.Pipeline;
 
@@ -17,11 +18,17 @@ namespace Azure.ResourceManager.Core
 
         public override void OnSendingRequest(HttpMessage message)
         {
-            var header = message.TryGetProperty("UserAgentOverride", out var userAgent) ? userAgent as string : _defaultHeader;
+            var header = message.TryGetProperty("SDKUserAgent", out var userAgent) ? userAgent as string : _defaultHeader;
             if (message.Request.Headers.TryGetValues(HttpHeader.Names.UserAgent, out var customHeaders))
             {
                 // append custom "user-agent" headers
-                header = $"{header} {string.Join(" ", customHeaders)}";
+                var strBuilder = new StringBuilder(header);
+                foreach (var customHeader in customHeaders)
+                {
+                    strBuilder.Append(' ');
+                    strBuilder.Append(customHeader);
+                }
+                header = strBuilder.ToString();
             }
             message.Request.Headers.SetValue(HttpHeader.Names.UserAgent, header);
         }