--- a/sdk/resourcemanager/Azure.ResourceManager/tests/Unit/MgmtTelemetryPolicyTests.cs
+++ b/sdk/resourcemanager/Azure.ResourceManager/tests/Unit/MgmtTelemetryPolicyTests.cs
@@ -25,7 +25,7 @@ namespace Azure.ResourceManager.Tests
             var message = new HttpMessage(requestMock, new ResponseClassifier());
             if (!string.IsNullOrEmpty(userAgentOverride))
             {
-                message.SetProperty("UserAgentOverride", userAgentOverride);
+                message.SetProperty("SDKUserAgent", userAgentOverride);
             }
             if (customHeaders.Length > 0)
             {