… sdk/resourcemanager/Azure.ResourceManager/src/Properties/AssemblyInfo.cs